refactor(outfitter): extract upgrade migration parsing/discovery helpers (OS-431)
