+++ apps/outfitter/package.json
@@ -195,6 +195,18 @@
         "default": "./dist/commands/upgrade-codemods.js"
       }
     },
+    "./commands/upgrade-migration-docs": {
+      "import": {
+        "types": "./dist/commands/upgrade-migration-docs.d.ts",
+        "default": "./dist/commands/upgrade-migration-docs.js"
+      }
+    },
+    "./commands/upgrade-migration-frontmatter": {
+      "import": {
+        "types": "./dist/commands/upgrade-migration-frontmatter.d.ts",
+        "default": "./dist/commands/upgrade-migration-frontmatter.js"
+      }
+    },
     "./commands/upgrade-planner": {
       "import": {
         "types": "./dist/commands/upgrade-planner.d.ts",
