+++ apps/outfitter/src/commands/upgrade-codemods.ts
@@ -14,7 +14,7 @@ import { isAbsolute, join, relative, resolve } from "node:path";
 import type { OutfitterError } from "@outfitter/contracts";
 import { InternalError, Result } from "@outfitter/contracts";
 
-import { readMigrationDocsWithMetadata } from "./upgrade.js";
+import { readMigrationDocsWithMetadata } from "./upgrade-migration-docs.js";
 
 // =============================================================================
 // Types
